refactor(backend): refactor code structure
