+++ lib/contentful/api.js
@@ -1,16 +1,7 @@
 import { remark } from "remark";
 import html from "remark-html";
 import remarkGfm from "remark-gfm";
-
-const POST_GRAPHQL_FIELDS = `
-    sys {
-        id
-    }  
-    title
-    date
-    excerpt
-    slug
-    tags`;
+import { POST_GRAPHQL_FIELDS } from "./config.js";
 
 /**
  * Connects to GraphQL API and returns all posts
@@ -20,7 +11,7 @@ const POST_GRAPHQL_FIELDS = `
  */
 async function fetchGraphQL(query, preview = false) {
   return fetch(
-    `https://graphql.contentful.com/content/v1/spaces/3se29gd38lrl/environments/master`,
+    `https://graphql.contentful.com/content/v1/spaces/${process.env.CONTENTFUL_SPACE_ID}/environments/${process.env.CONTENTFUL_ENVIRONMENT}`,
     {
       method: "POST",
       headers: {
@@ -39,9 +30,9 @@ async function fetchGraphQL(query, preview = false) {
 /**
  * Query to get all posts
  * @param {Boolean} preview
- * @return {Array} Posts Data
+ * @return {Array} Posts Data Entries
  */
-export async function getAllPostsForHome(preview) {
+export async function getAllPosts(preview) {
   const entries = await fetchGraphQL(
     `query {
         blogPostCollection(order: date_DESC, preview: ${
@@ -54,32 +45,16 @@ export async function getAllPostsForHome(preview) {
         }`,
     preview
   );
-  return extractPostEntries(entries);
-}
 
-/**
- * Fetching all posts
- * @param {JSON} fetchResponse
- * @return {JSON} Posts Entries
- */
-function extractPostEntries(fetchResponse) {
-  return fetchResponse?.data?.blogPostCollection?.items;
-}
-
-/**
- * Fetching a single post
- * @param {JSON} fetchResponse
- * @return {JSON} Posts Entries
- */
-function extractPost(fetchResponse) {
-  return fetchResponse?.data?.blogPost;
+  const AllPostData = entries.data?.blogPostCollection?.items;
+  return AllPostData;
 }
 
 /**
  * Query a single post by id
- * @param {boolean} preview
+ * @param {Boolean} preview
  * @param {String} postId
- * @return {JSON} Single post data
+ * @return {Object} Single post data
  */
 export async function getPost(preview, postId) {
   // Get a singlepost/entry
@@ -93,7 +68,7 @@ export async function getPost(preview, postId) {
         }`
   );
 
-  const postData = extractPost(entry);
+  const postData = entry.data?.blogPost;
 
   // Use remark to convert markdown into HTML string
   const processedContent = await remark()
@@ -101,6 +76,7 @@ export async function getPost(preview, postId) {
     .use(remarkGfm)
     .process(postData.excerpt);
   const contentHtml = processedContent.toString();
+
   return {
     postData: {
       ...postData,
