+++ lib/contentful/config.js
@@ -1,0 +1,9 @@
+export const POST_GRAPHQL_FIELDS = `
+    sys {
+        id
+    }  
+    title
+    date
+    excerpt
+    slug
+    tags`;
